// src/quotation/domain/enums/BookingStatus.ts
export enum BookingStatus {
    DRAFT = 'DRAFT',
    CONFIRMED = 'CONFIRMED',
    AWAITING_PAYMENT = 'AWAITING_PAYMENT',
    PAYMENT_PROCESSING = 'PAYMENT_PROCESSING',
    PAYMENT_FAILED = 'PAYMENT_FAILED',
    PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
    CANCELED = 'CANCELED',
    COMPLETED = 'COMPLETED'
} 